move MemoryMeasurements tests to single file
